Ensure outdated packages get removed from the cache
If a package has a new version released upstream the prior version
will remain in the cache. Similarly if a package has any
dependencies changed to a different package or removed, those
unneeded packages will remain in the cache as well. Due to this
the slug size will slowly balloon overtime as long as the stack
or Aptfile is not changed.

This tracks the packages that are downloaded during installation
and removes all debs not in that list from the cache archive.
`apt-get autoclean` is not used because it will only clean out
packages that can no longer be downloaded so not all  outdated
versions will be removed. If the cache has version 1, 2, and 3
of a package in it and version 1 is removed from upstream
then version 1 will be removed but 2 will remain.