Docs: Make the example more obviously an example (#80)
In a recent support ticket, a user had copied the example `:repo:` line into their `Aptfile`,
which previously referred to the now obsolete "artful" Ubuntu LTS release.

This caused the build to fail after #79, when previously the error was ignored.

This PR makes the example more clearly just an example, and emphasises that
using a custom repository is only necessary if using packages from outside
the standard Ubuntu repository.

In addition, the bogus `pg` Gem example was removed, since `libpq-dev` is
part of the stack image, so we should not be encouraging people to install
it unnecessarily.