Empty Aptfile fix
This PR fixes an error that occurs if an Aptfile with no packages is used. This case is detected now and a warning is printed advising that the buildpack and Aptfile should be removed before exiting the build.

I've also added a test harness to this project with several variations of Aptfile contents we've seen used. There appears to have been tests in this repository at an earlier point but they were lost or abandoned at some point. This should prevent future accidental breakage if changes are needed for this buildpack.